Services, types and static implementations
