--- src/app/tab2/productcard.page.ts
@@ -1,18 +1,0 @@
-import ProductCard from "./types/productcard.type"
-
-export let productcard: ProductCard[] = [
-    {
-        alt: 'Silhouette of mountains',
-        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
-        name: 'Name product',
-        quantity: 'Quantities',
-        category: 'Categories'
-    },
-    {
-        alt: 'Silhouette of mountains',
-        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
-        name: 'Name product',
-        quantity: 'Quantities',
-        category: 'Categories',
-    }
-]
