+++ src/app/tab2/static/productcard.page.ts
@@ -1,0 +1,39 @@
+import ProductCard from "../types/productcard.type"
+
+export let productcard: ProductCard[] = [
+    {
+        alt: 'Silhouette of mountains',
+        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
+        name: 'Name product',
+        quantity: 'Quantities',
+        category: 'Categories'
+    },
+    {
+        alt: 'Silhouette of mountains',
+        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
+        name: 'Name product',
+        quantity: 'Quantities',
+        category: 'Categories'
+    },
+    {
+        alt: 'Silhouette of mountains',
+        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
+        name: 'Name product',
+        quantity: 'Quantities',
+        category: 'Categories'
+    },
+    {
+        alt: 'Silhouette of mountains',
+        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
+        name: 'Name product',
+        quantity: 'Quantities',
+        category: 'Categories'
+    },
+    {
+        alt: 'Silhouette of mountains',
+        src: 'https://ionicframework.com/docs/img/demos/thumbnail.svg',
+        name: 'Name product',
+        quantity: 'Quantities',
+        category: 'Categories',
+    }
+]
