+++ src/app/tab2/tab2.page.html
@@ -41,4 +41,4 @@
       </ion-list>
     </ion-card-content>
   </ion-card>
-  <ion-content />
+</ion-content>
