+++ src/app/tab2/types/itemcards.type.ts
@@ -1,0 +1,4 @@
+import ProductCard from "./productcard.type"
+import PubCards from "./pubcards.type"
+
+type ItemCard = ProductCard | PubCards
