+++ src/app/tab2/types/pubcards.type.ts
@@ -1,8 +1,8 @@
 type PubCards = {
-    alt: string;
-    src: string;
-    name: string;
-    description: string;
+    alt: string
+    src: string
+    name: string
+    description: string
 }
 
-export default PubCards;
+export default PubCards
