+++ src/services/pub/pub.service.ts
@@ -1,3 +1,4 @@
+import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { PubInterfaceService } from './interfaces/pub.interface.service';
 
@@ -6,5 +7,9 @@ import { PubInterfaceService } from './interfaces/pub.interface.service';
 })
 
 export class PubService implements PubInterfaceService {
-  constructor() { }
+  private apiBaseUrl = 'https://world.openfoodfacts.org/api/v0/product/';
+
+  constructor(private http: HttpClient) {
+
+  }
 }
